Fix print non-raw markup when stdout is not a TTY
This commit fixes a bug where non-raw markup was printed when stdout was not a TTY. Now, if stdout is not a TTY, the raw flag is used to determine whether or not to print raw or formatted text. The function `update_markdown_stream` has been updated to use `sys.stdout.isatty()` to determine if stdout is a TTY. The unnecessary whitespace has been removed. The function `get_template_content` has not been modified.